server/requirements.txt
openenv-core>=0.2.1
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=2.30.0
